app/requirements.txt
apscheduler >= 3.10
pytz >= 2023
pyyaml >= 6.0
hydroloader >= 0.1.0